Test use of media ranges in `accept` arg of add_view and add_route.
These tests are the same as those in the 'accept-arg-current-behavior'
branch.

The four tests with errors will pass if we comment out line 199 in
WebOb's acceptparse.py (the call to _check_offer()). The calls to
_check_offer() were removed in the GSOC changes, because _check_offer()
in preventing the use of media ranges was inconsistent with both the
documentation and the behavior when the Accept header was valid -- it
appeared to be an oversight from the re-use of NilAccept across the
different Accept* headers.